+++ src/settings/MappingProfiles/MappingProfilesForm.js
@@ -49,6 +49,7 @@ import {
   MappingItemDetails,
   MappingMARCBibDetails,
   MappingInvoiceDetails,
+  MappingMARCAuthorityDetails,
 } from './detailsSections/edit';
 
 import {
@@ -273,6 +274,7 @@ export const MappingProfilesFormComponent = ({
     fieldMappingsForMARCField: fieldMappingsForMARC,
     onUpdateFieldAdd: setInitialValuesForMARCUpdates,
     setReferenceTables: setFormFieldValue,
+    folioRecordType,
   };
 
   const invoiceDetailsProps = {
@@ -295,6 +297,7 @@ export const MappingProfilesFormComponent = ({
     HOLDINGS: <MappingHoldingsDetails {...detailsProps} />,
     ITEM: <MappingItemDetails {...detailsProps} />,
     MARC_BIBLIOGRAPHIC: <MappingMARCBibDetails {...MARCDetailsProps} />,
+    MARC_AUTHORITY: <MappingMARCAuthorityDetails {...MARCDetailsProps} />,
     INVOICE: <MappingInvoiceDetails {...invoiceDetailsProps} />,
   };
 
@@ -394,7 +397,7 @@ export const MappingProfilesFormComponent = ({
                   />
                 </div>
               </Accordion>
-              {folioRecordType !== MARC_TYPES.MARC_AUTHORITY && (
+              {folioRecordType && (
                 <Accordion
                   id="mapping-profile-details"
                   label={<FormattedMessage id="ui-data-import.details" />}