+++ src/settings/MappingProfiles/ViewMappingProfile/ViewMappingProfile.js
@@ -262,6 +262,7 @@ export class ViewMappingProfile extends Component {
       marcMappingOption,
       marcFieldProtectionFields,
       mappingMarcFieldProtectionFields,
+      folioRecordType: existingRecordType,
     };
 
     const mappingFields = mappingDetails?.mappingFields || [];